Updating xo and ava. Renaming plugin files. Adding extra option to remove critical css rules from origin file.
