+++ .editorconfig
@@ -1,7 +1,7 @@
 root = true
 
 [*]
-indent_style = space
+indent_style = tab
 end_of_line = lf
 charset = utf-8
 trim_trailing_whitespace = true
